r e f a c t o r

+++ src/main/kotlin/dev/yeseong0412/authtemplate/domain/chat/service/ChatRoomServiceImpl.kt
@@ -9,8 +9,10 @@ import dev.yeseong0412.authtemplate.domain.chat.domain.repository.ChatMessageRep
 import dev.yeseong0412.authtemplate.domain.chat.exception.ChatRoomErrorCode
 import dev.yeseong0412.authtemplate.domain.chat.presentation.dto.request.ChatMessage
 import dev.yeseong0412.authtemplate.domain.chat.presentation.dto.response.ChatOnline
+import dev.yeseong0412.authtemplate.domain.chat.presentation.dto.response.ChatRoom
 import dev.yeseong0412.authtemplate.domain.user.domain.repository.UserRepository
 import dev.yeseong0412.authtemplate.domain.user.exception.UserErrorCode
+import dev.yeseong0412.authtemplate.domain.user.presentation.dto.response.ParticipantsInfo
 import dev.yeseong0412.authtemplate.global.security.jwt.util.JwtUtils
 import dev.yeseong0412.authtemplate.global.common.BaseResponse
 import dev.yeseong0412.authtemplate.global.exception.CustomException
@@ -29,18 +31,18 @@ class ChatRoomServiceImpl(
 ) : ChatRoomService {
 
     @Transactional(readOnly = true)
-    override fun getAllRooms(): BaseResponse<List<ChatRoomInfo>> {
+    override fun getAllRooms(): BaseResponse<List<ChatRoom>> {
         val rooms = chatRoomRepository.findAll()
         return BaseResponse(
             message = "success",
-            data = rooms.map { ChatRoomInfo(it.id, it.name, it.participants.map { pr -> pr.name }) }
+            data = rooms.map { ChatRoom(id = it.id, name = it.name, participants = it.participants.map { pr -> pr.name }) }
         )
     }
 
     @Transactional
     override fun createRoom(name: String, userId: Long): BaseResponse<Unit> {
         val user = userRepository.findById(userId).orElseThrow { CustomException(UserErrorCode.USER_NOT_FOUND) }
-        val room = ChatRoomEntity(name = name, participants = mutableSetOf(user))
+        val room = ChatRoomEntity(name = name, participants = mutableSetOf(user), adminId = userId)
 
         chatRoomRepository.save(room)
         user.rooms.add(room)
@@ -53,8 +55,7 @@ class ChatRoomServiceImpl(
 
     @Transactional
     override fun inviteToRoom(roomId: Long, userEmail: String): BaseResponse<Unit> {
-        val room =
-            chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
+        val room = chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
         val user = userRepository.findByEmail(userEmail) ?: throw CustomException(UserErrorCode.USER_NOT_FOUND)
 
         if (room.participants.size >= 8) throw CustomException(ChatRoomErrorCode.CHAT_ROOM_NUMBER_LIMIT_EXCEEDED)
@@ -70,12 +71,31 @@ class ChatRoomServiceImpl(
     }
 
     @Transactional
-    override fun deleteRoom(roomId: Long): BaseResponse<Unit> {
-        val room =
-            chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
+    override fun leaveRoom(roomId: Long, userId: Long): BaseResponse<Unit> {
+        val room = chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
+        val user = userRepository.findById(userId).orElseThrow { CustomException(UserErrorCode.USER_NOT_FOUND) }
+
+        if (userId == room.adminId) deleteRoom(roomId, userId)
+
+        room.participants.remove(user)
+        user.rooms.remove(room)
+        chatRoomRepository.save(room)
+        userRepository.save(user)
+
+        if (room.participants.size == 0) chatRoomRepository.delete(room)
+
+        return BaseResponse(
+            message = "success"
+        )
+    }
+
+    @Transactional
+    override fun deleteRoom(roomId: Long, userId: Long): BaseResponse<Unit> {
+        val room = chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
+
+        if (room.adminId != userId) throw CustomException(ChatRoomErrorCode.CANNOT_DELETE_CHATROOM)
 
         room.participants.forEach { user -> user.rooms.remove(room) }
-
         chatRoomRepository.delete(room)
 
         return BaseResponse(
@@ -86,7 +106,7 @@ class ChatRoomServiceImpl(
     @Transactional(readOnly = true)
     override fun getRoomInfo(roomId: Long): BaseResponse<ChatRoomInfo> {
         val roomInfo = chatRoomRepository.findById(roomId)
-            .map { ChatRoomInfo(id = it.id, name = it.name, participants = it.participants.map { pr -> pr.name }) }
+            .map { ChatRoomInfo(id = it.id, name = it.name, participants = it.participants.map { pr -> ParticipantsInfo(id = pr.id, email = pr.email, name = pr.name) }) }
             .orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
 
         return BaseResponse(
@@ -97,23 +117,16 @@ class ChatRoomServiceImpl(
 
     @Transactional(readOnly = true)
     override fun enterRoom(roomId: Long, userId: Long): ChatOnline {
-        val room =
-            chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
+        val room = chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
         val user = userRepository.findById(userId).orElseThrow { CustomException(UserErrorCode.USER_NOT_FOUND) }
 
         return ChatOnline(writer = "시스템", message = "${user.name} 님이 ${room.name}에 입장하셨습니다.")
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     override fun exitRoom(roomId: Long, userId: Long): ChatOnline {
-        val room =
-            chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
+        val room = chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
         val user = userRepository.findById(userId).orElseThrow { CustomException(UserErrorCode.USER_NOT_FOUND) }
-
-        room.participants.remove(user)
-        chatRoomRepository.save(room)
-
-        if (room.participants.size == 0) chatRoomRepository.delete(room)
 
         return ChatOnline(writer = "시스템", message = "${user.name} 님이 ${room.name}에서 퇴장하셨습니다.")
     }
@@ -130,8 +143,7 @@ class ChatRoomServiceImpl(
 
     @Transactional(readOnly = true)
     override fun getAllMessages(roomId: Long, objectId: String?, userId: Long): BaseResponse<List<ChatMessageInfo>> {
-        val room =
-            chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
+        val room = chatRoomRepository.findById(roomId).orElseThrow { CustomException(ChatRoomErrorCode.CHAT_ROOM_NOT_FOUND) }
         val id = if (objectId != null) ObjectId(objectId) else ObjectId()
         val pageable = PageRequest.of(0, 20, Sort.by(Sort.Direction.DESC, "_id"))
         val messages = chatMessageRepository.findMessagesByRoomIdAndObjectId(roomId, id, pageable)
@@ -143,7 +155,7 @@ class ChatRoomServiceImpl(
                         .orElseThrow { CustomException(UserErrorCode.USER_NOT_FOUND) }.name,
                     content = it.content,
                     isMine = it.writerId == userId,
-                    timestamp = it.timestamp
+                    timestamp = it.timestamp.toString()
                 )
             }
